+++ integration.py
@@ -16,13 +16,13 @@ class Domain:
 
 def sample(x1, x2, n):
     d = x1.shape[0]
-    U = tf.random.uniform(shape=(n,d), dtype=tf.float64)
+    U = tf.random.uniform(shape=(n,d), dtype=tf.float32)
     x = U * (x2 - x1) + x1
     return x
 
 
 def mc_int(G, f, n):
-    sides = tf.abs(G.x2 - G.x1)
+    sides = tf.cast(tf.abs(G.x2 - G.x1),dtype=tf.float32)
     vol = tf.reduce_prod(sides)
     x = sample(G.x1, G.x2, n)
     mask = G.indicator(x)
@@ -34,7 +34,7 @@ def mc_int(G, f, n):
 
 def check_boundary_cond(f1, f2, G, n):
     x = tf.random.uniform(shape=(n,),
-                          minval=G.a, maxval=G.b, dtype=tf.float64)
+                          minval=G.a, maxval=G.b, dtype=tf.float32)
     path = G.gamma(x)
     delta = f1(path) - f2(path)
     return (G.b - G.a) * tf.reduce_mean(delta**2)
